requirements.txt
DBUtils==3.1.0
Flask==3.0.3
numpy==1.26.4
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
Pillow~=10.2.0
PyMySQL==1.1.1
torch==2.2.2
gevent==24.2.1
psycopg2==2.9.9
yaml~=0.2.5
pyyaml~=6.0.1